+++ lib/ref_inspector/downloader.ex
@@ -12,6 +12,7 @@ defmodule RefInspector.Downloader do
   """
   @spec download() :: :ok
   def download() do
+    Application.ensure_all_started(:hackney)
     File.mkdir_p!(Config.database_path())
 
     Enum.each(Config.yaml_urls(), fn config ->