+++ lib/ref_inspector/database/loader.ex
@@ -1,6 +1,8 @@
 defmodule RefInspector.Database.Loader do
   @moduledoc false
 
+  alias RefInspector.Config
+
   @doc """
   Returns the yaml contents of a database file.
   """
@@ -8,8 +10,9 @@ defmodule RefInspector.Database.Loader do
   def load(file) do
     case File.regular?(file) do
       true ->
-        file
-        |> :yamerl_constr.file([:str_node_as_binary])
+        {reader_mod, reader_fun, reader_extra_args} = Config.yaml_file_reader()
+
+        apply(reader_mod, reader_fun, [file | reader_extra_args])
         |> maybe_hd()
 
       false ->