+++ CHANGELOG.md
@@ -52,11 +52,11 @@
 
 - Enhancements
     - All databases can be reloaded (asynchronously) using `RefInspector.reload_databases/0`
-    - Configuration can be done on supervisor (re-) start by setting a `{ mod, fun }` tuple for the config key `:init`. This method will be called without arguments
+    - Configuration can be done on supervisor (re-) start by setting a `{mod, fun}` tuple for the config key `:init`. This method will be called without arguments
     - When using the mix download task with a default remote configuration an information README file is placed next to the downloaded file(s)
 
 - Soft deprecations (no warnings)
-    - Support for `{ :system, "ENV_VARIABLE" }` configuration has been removed from the documentation. It will eventually be removed completely after a proper deprecation phase
+    - Support for `{:system, "ENV_VARIABLE"}` configuration has been removed from the documentation. It will eventually be removed completely after a proper deprecation phase
 
 ## v0.16.0 (2017-09-24)